requirements.txt
streamlit
requests
pytz
watchdog
torch
transformers
opencv-python
Pillow
timm